Accesslog 5

+++ src/main/java/csc369/AccessLog8.java
@@ -4,33 +4,33 @@ import java.io.IOException;
 import java.util.Iterator;
 
 import org.apache.hadoop.io.Text;
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.mapreduce.Mapper;
 import org.apache.hadoop.mapreduce.Reducer;
 
 public class AccessLog8 {
 
-    public static final Class OUTPUT_KEY_CLASS = LongWritable.class;
+    public static final Class OUTPUT_KEY_CLASS = IntWritable.class;
     public static final Class OUTPUT_VALUE_CLASS = Text.class;
 
-    public static class MapperImpl extends Mapper<LongWritable, Text, LongWritable, Text> {
+    public static class MapperImpl extends Mapper<IntWritable, Text, IntWritable, Text> {
 
         @Override
-	protected void map(LongWritable key, Text value,
+	protected void map(IntWritable key, Text value,
 			   Context context) throws IOException, InterruptedException {
 	    String[] sa = value.toString().split(" ");
 	    Text hostname = new Text();
         hostname.set(sa[0]);
-        LongWritable count = new LongWritable(Long.parseLong(sa[1]));
+        IntWritable count = new IntWritable(Integer.parseInt(sa[1]));
 	    context.write(count, hostname);
         }
     }
 
     
-    public static class ReducerImpl extends Reducer<LongWritable, Text, Text, LongWritable> {
+    public static class ReducerImpl extends Reducer<IntWritable, Text, Text, IntWritable> {
 	
         @Override
-	protected void reduce(LongWritable count, Iterable<Text> hostnames,
+	protected void reduce(IntWritable count, Iterable<Text> hostnames,
 			      Context context) throws IOException, InterruptedException {
             
             Iterator<Text> itr = hostnames.iterator();
